+++ tests/infrastructures/Symfony/Components/Form/Type/ClusterCredentialsTypeTest.php
@@ -25,8 +25,12 @@ declare(strict_types=1);
 
 namespace Teknoo\Tests\East\Paas\Infrastructures\Symfony\Form\Type;
 
+use ArrayIterator;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Form\AbstractType;
+use Symfony\Component\Form\DataMapperInterface;
+use Symfony\Component\Form\FormBuilderInterface;
+use Symfony\Component\Form\FormInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 use Teknoo\East\Paas\Infrastructures\Symfony\Form\Type\ClusterCredentialsType;
 use Teknoo\East\Paas\Object\ClusterCredentials;
@@ -53,13 +57,13 @@ class ClusterCredentialsTypeTest extends TestCase
     private function getFormArray(): array
     {
         return [
-            'name' => 'fooBar',
             'caCertificate' => 'babar',
             'clientCertificate' => 'babar',
             'clientKey' => 'babar',
             'token' => 'bar',
             'username' => 'foo',
             'password' => 'fooBar',
+            'clear' => false,
         ];
     }
 
@@ -70,6 +74,52 @@ class ClusterCredentialsTypeTest extends TestCase
             $this->buildForm()->configureOptions(
                 $this->createMock(OptionsResolver::class)
             )
+        );
+    }
+
+    public function testDataMapperWithSecureData()
+    {
+        $builder = $this->createMock(FormBuilderInterface::class);
+
+        $object = new ClusterCredentials(
+            caCertificate: 'babar',
+            clientCertificate: 'babar',
+            clientKey: 'babar',
+            token: 'bar',
+            username: 'foo',
+            password: 'fooBar',
+        );
+
+        $builder->expects(self::any())
+            ->method('setDataMapper')
+            ->willReturnCallback(function (DataMapperInterface $dataMapper) use ($builder, $object) {
+                $children = [];
+                $formArray = $this->getFormArray();
+                $formArray['password'] = '';
+                $formArray['token'] = '';
+                $formArray['clear'] = false;
+                foreach ($formArray as $name=>$value) {
+                    $mock = $this->createMock(FormInterface::class);
+                    $mock->expects(self::once())->method('getData')->willReturn($value);
+                    $children[$name] = $mock;
+                }
+
+                $forms = new ArrayIterator($children);
+                $dataMapper->mapDataToForms($object, $forms);
+                $object2 = '';
+                $dataMapper->mapFormsToData($forms, $object2);
+
+                self::assertEquals(
+                    $object,
+                    $object2,
+                );
+
+                return $builder;
+            });
+
+        self::assertInstanceOf(
+            AbstractType::class,
+            $this->buildForm()->buildForm($builder, [])
         );
     }
 }