+++ tests/infrastructures/Symfony/Components/Form/Type/FormTestTrait.php
@@ -26,7 +26,6 @@ declare(strict_types=1);
 namespace Teknoo\Tests\East\Paas\Infrastructures\Symfony\Form\Type;
 
 use Doctrine\Bundle\MongoDBBundle\Form\Type\DocumentType;
-use Doctrine\MongoDB\Query\Builder;
 use Doctrine\ODM\MongoDB\Repository\DocumentRepository;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\DataMapperInterface;
@@ -60,34 +59,6 @@ trait FormTestTrait
     public function testBuildForm()
     {
         $builder = $this->createMock(FormBuilderInterface::class);
-
-        $builder->expects(self::any())
-            ->method('add')
-            ->willReturnCallback(
-                function ($child, $type, array $options = array()) use ($builder) {
-                    if (DocumentType::class == $type && isset($options['query_builder'])) {
-                        $qBuilder = $this->createMock(Builder::class);
-                        $qBuilder->expects(self::once())
-                            ->method('field')
-                            ->with('deletedAt')
-                            ->willReturnSelf();
-
-                        $qBuilder->expects(self::once())
-                            ->method('equals')
-                            ->with(null)
-                            ->willReturnSelf();
-
-                        $repository = $this->createMock(DocumentRepository::class);
-                        $repository->expects(self::once())
-                            ->method('createQueryBuilder')
-                            ->willReturn($qBuilder);
-
-                        $options['query_builder']($repository);
-                    }
-
-                    return $builder;
-                }
-            );
 
         $builder->expects(self::any())
             ->method('addEventListener')