+++ tests/infrastructures/Symfony/Components/Form/Type/SshIdentityTypeTest.php
@@ -25,8 +25,12 @@ declare(strict_types=1);
 
 namespace Teknoo\Tests\East\Paas\Infrastructures\Symfony\Form\Type;
 
+use ArrayIterator;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Form\AbstractType;
+use Symfony\Component\Form\DataMapperInterface;
+use Symfony\Component\Form\FormBuilderInterface;
+use Symfony\Component\Form\FormInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 use Teknoo\East\Paas\Infrastructures\Symfony\Form\Type\SshIdentityType;
 use Teknoo\East\Paas\Object\SshIdentity;
@@ -55,6 +59,7 @@ class SshIdentityTypeTest extends TestCase
         return [
             'name' => 'fooBar',
             'privateKey' => 'barFoo',
+            'clear' => false,
         ];
     }
 
@@ -65,6 +70,47 @@ class SshIdentityTypeTest extends TestCase
             $this->buildForm()->configureOptions(
                 $this->createMock(OptionsResolver::class)
             )
+        );
+    }
+
+    public function testDataMapperWithSecureData()
+    {
+        $builder = $this->createMock(FormBuilderInterface::class);
+
+        $object = new SshIdentity(
+            name: 'fooBar',
+            privateKey: 'barFoo',
+        );
+
+        $builder->expects(self::any())
+            ->method('setDataMapper')
+            ->willReturnCallback(function (DataMapperInterface $dataMapper) use ($builder, $object) {
+                $children = [];
+                $formArray = $this->getFormArray();
+                $formArray['privateKey'] = '';
+                $formArray['clear'] = false;
+                foreach ($formArray as $name=>$value) {
+                    $mock = $this->createMock(FormInterface::class);
+                    $mock->expects(self::once())->method('getData')->willReturn($value);
+                    $children[$name] = $mock;
+                }
+
+                $forms = new ArrayIterator($children);
+                $dataMapper->mapDataToForms($object, $forms);
+                $object2 = '';
+                $dataMapper->mapFormsToData($forms, $object2);
+
+                self::assertEquals(
+                    $object,
+                    $object2,
+                );
+
+                return $builder;
+            });
+
+        self::assertInstanceOf(
+            AbstractType::class,
+            $this->buildForm()->buildForm($builder, [])
         );
     }
 }